Refactor fetch and prompt context functions, update GitClient and GitHubClient, and add nested prompt sample script
